+++ data/out.txt
@@ -1,0 +1,2 @@
+What in the world
+Is up
